appli/requirements.txt
aniso8601==9.0.1
cffi==1.14.5
click==7.1.2
cryptography==3.4.6
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
flask-swagger==0.2.14
flask-swagger-ui==3.36.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==2.0.1
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
Werkzeug==1.0.1